requirements.txt
streamlit
gunicorn
Flask
numpy
keras == 2.13.1
tensorflow == 2.13.0
Pillow == 9.4.0